requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain-groq
langchain-core
langgraph
langchain-community
arxiv
wikipedia
tavily-python